Added tests for boost odeint ABM wrapper
